+++ Shared/Dictionary/DictionaryContentView.swift
@@ -110,7 +110,7 @@ struct DictionaryContentView_Previews: PreviewProvider {
     static let soundService = SoundService()
     static let userDefaultsStore = UserDefaultsStore()
     static let favoritesService = PhraseFavoritesService(userDefaultsStore: userDefaultsStore,
-                                                         dictionaryDataStore: DictionaryDataStore())
+                                                         dictionaryDataStore: DictionaryDataStore.shared)
     static let favoritesProvider = PhrasesFavoritesProvider(favoritesService: favoritesService)
 
     static let phrases: [Dictionary.PhraseID: Dictionary.Phrase] = [